refactor:  Add WalletEx class to handle coinjoin related functionality
